requirements.txt
# Logic Defense System 2.0 - Dependencies

# Core
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Crawler
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
supabase>=2.3.0
psycopg2-binary>=2.9.0
pgvector>=0.2.4

# LangChain & RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# AI & ML
openai>=1.10.0
anthropic>=0.18.0
tenacity>=8.2.0
tiktoken>=0.5.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Vector Search & Embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.22

# Telegram Bot
python-telegram-bot>=20.7
python-telegram-bot[job-queue]>=20.7

# Utilities
colorlog>=6.7.0
pydantic>=2.5.0
croniter>=2.0.0

# Monitoring
prometheus-client>=0.19.0
psutil>=5.9.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
pylint>=3.0.0
mypy>=1.7.0
